+++ src/containers/Home.js
@@ -13,7 +13,6 @@ import twreporterRedux from '@twreporter/redux'
 import { SITE_NAME, SITE_META } from '../constants/index'
 import { connect } from 'react-redux'
 import { signOutAction } from '@twreporter/registration'
-import PropTypes from 'prop-types'
 
 // lodash
 import get from 'lodash/get'